improvements: mouse coordinate works even with view/camera is set

+++ modules/event/antara/gaming/event/mouse.button.pressed.cpp
@@ -18,7 +18,7 @@
 
 namespace antara::gaming::event
 {
-    mouse_button_pressed::mouse_button_pressed(input::mouse_button button_, int x_, int y_) noexcept
+    mouse_button_pressed::mouse_button_pressed(input::mouse_button button_, float x_, float y_) noexcept
             : button(button_), x(x_), y(y_)
     {
     }
